+++ packages/sdk-router/src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './bridge'
 export * from './intent'
 export * from './misc'
+export * from './swap'
